requirements.txt
cloudscraper
coloredlogs
flask
flask-cors
minio
python-dotenv
redis
requests
schedule
